fix(workspaces): keep selected quotes uncited
