+++ src/features/workspaces/model/workspace-ai-context-types.ts
@@ -1,7 +1,6 @@
 import type { WorkspaceTab } from "#/features/workspaces/model/tab-types";
 import type { WorkspaceItem } from "#/features/workspaces/model/types";
 import type { WorkspaceItemFacts } from "#/features/workspaces/contracts";
-import type { WorkspaceReferenceRecord } from "#/features/workspaces/ai/workspace-reference";
 import type {
 	WorkspaceAiContextItemViewState,
 	WorkspaceItemViewState,
@@ -63,7 +62,6 @@ export type WorkspaceAiContextOutlineItem = {
 };
 
 export type WorkspaceAiContextSnapshotSelectedQuote = {
-	citation?: WorkspaceReferenceRecord;
 	label: string;
 	order: number;
 	source:
